+++ modules/culling/test/index.ts
@@ -13,6 +13,7 @@ import './lib/bounding-volumes/oriented-bounding-box.spec';
 
 import './lib/plane.spec';
 import './lib/ray.spec';
+import './lib/shapes.spec';
 import './lib/culling-volume.spec';
 
 import './lib/perspective-off-center-frustum.spec';